% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Binary.R
\name{Run_Binary}
\alias{Run_Binary}
\title{Binary logistic IRLS-SuSiE path}
\usage{
Run_Binary(
  X,
  y,
  Z = NULL,
  family = binomial(link = "logit"),
  weight_cutoff = 0.005,
  L,
  max.iter,
  min.iter,
  max.eps,
  susie.iter,
  verbose = TRUE,
  n_threads = 1,
  coverage = 0.9,
  estimate_residual_variance = FALSE,
  scaled_prior_variance = 1,
  residual_variance = 1,
  L.init = 1,
  init_cor_method = c("pearson", "spearman"),
  ...
)
}
\description{
Binary logistic IRLS-SuSiE path
}
