% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_Cox.R
\name{Run_Cox}
\alias{Run_Cox}
\title{Cox score IRLS-SuSiE path}
\usage{
Run_Cox(
  X,
  y,
  status,
  Z = NULL,
  L,
  max.iter,
  min.iter,
  max.eps,
  susie.iter,
  verbose = TRUE,
  n_threads = 1,
  coverage = 0.9,
  estimate_residual_variance = FALSE,
  scaled_prior_variance = 1,
  residual_variance = 1,
  ridge = 1e-06,
  L.init = 1,
  init_cor_method = c("pearson", "spearman"),
  ...
)
}
\description{
Cox score IRLS-SuSiE path
}
