% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_GLM.R
\name{Run_GLM}
\alias{Run_GLM}
\title{General GLM IRLS-SuSiE path}
\usage{
Run_GLM(
  X,
  y,
  Z = NULL,
  weight_cutoff = 0.005,
  family = binomial(link = "logit"),
  L,
  max.iter,
  min.iter,
  max.eps,
  susie.iter,
  verbose = TRUE,
  n_threads = 1,
  coverage = 0.9,
  estimate_residual_variance = FALSE,
  residual_variance = 1,
  scaled_prior_variance = 1,
  L.init = 1,
  init_cor_method = c("pearson", "spearman"),
  ...
)
}
\description{
General GLM IRLS-SuSiE path
}
