% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Run_NB.R
\name{Run_NB}
\alias{Run_NB}
\title{Negative-binomial IRLS-SuSiE path}
\usage{
Run_NB(
  X,
  y,
  Z = NULL,
  weight_cutoff = 0.005,
  L,
  max.iter,
  min.iter,
  max.eps,
  susie.iter,
  verbose = TRUE,
  n_threads = 1,
  coverage = 0.9,
  estimate_residual_variance = FALSE,
  residual_variance = 1,
  scaled_prior_variance = 1,
  theta_init = 10,
  estimate_theta = TRUE,
  L.init = 1,
  init_cor_method = c("pearson", "spearman"),
  ...
)
}
\description{
Negative-binomial IRLS-SuSiE path
}
