% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuSiEIRLS-package.R
\docType{package}
\name{SuSiEIRLS-package}
\alias{SuSiEIRLS}
\alias{SuSiEIRLS-package}
\title{SuSiEIRLS: SuSiE for non-Gaussian outcomes via IRLS}
\description{
Implements SuSiE model for extended generalized linear models (EGLMs) using an Iteratively Reweighted Least Squares (IRLS) algorithm, and for Cox proportional-hazards survival outcomes via score-based sufficient statistics, and ordered-categorical outcomes via cumulative-logit score statistics. The package extends the Sum of Single Effects (SuSiE) framework to logistic, negative binomial, and other GLMs, as well as time-to-event data, fine-mapping non-Gaussian ordered, and censored outcomes within a unified iterative framework.
}
\keyword{internal}
